Extract usePushNotifications from App.tsx
Move the device-scoped push notification setup — the Capacitor branch
(register, listeners, permission flow) and the web/PWA branch
(standalone-mode-gated permission request + onForegroundMessage
subscription + 8-loop alert sound) — into usePushNotifications().

Drops the @capacitor/push-notifications, @capacitor/core, and
onForegroundMessage imports from App.tsx. The audioRef moves into the
hook; useNag continues to maintain its own independent audioRef on
purpose so the nag loop and per-push alert never share an element.

The hook returns { fcmToken, setFcmToken } so the existing
auto-clock-in effect (which mirrors the token into
bars/{barId}/tokens/{uid}) keeps working unchanged, and the PWA
install-prompt onInstalled callback can still re-request permission
on accept.

53/53 tests pass.
